credentials issuance examples
