+++ .gitignore
@@ -38,10 +38,41 @@ examples/client/simple_chat/simple_chat
 examples/client/simple_chat/simple_chat_storage
 examples/client/advanced_features/advanced_features
 examples/client/credential_issuance/credential_issuance
+examples/client/credential_issuance/basic/basic
+examples/client/credential_issuance/multi_claim/multi_claim
+examples/client/credential_issuance/evidence/evidence
+examples/client/credential_issuance/complex/complex
+examples/client/credential_issuance/advanced/advanced
 examples/client/credential_issuance/issuer_storage
 examples/client/credential_issuance/holder_storage
+examples/client/credential_issuance/simple_issuer_storage
+examples/client/credential_issuance/simple_holder_storage
+examples/client/credential_issuance/basic/basic_issuer_storage
+examples/client/credential_issuance/basic/basic_holder_storage
+examples/client/credential_issuance/multi_claim/multi_issuer_storage
+examples/client/credential_issuance/multi_claim/multi_holder_storage
+examples/client/credential_issuance/evidence/evidence_issuer_storage
+examples/client/credential_issuance/evidence/evidence_holder_storage
+examples/client/credential_issuance/complex/complex_issuer_storage
+examples/client/credential_issuance/complex/complex_holder_storage
+examples/client/credential_issuance/advanced/issuer_storage
+examples/client/credential_issuance/advanced/holder_storage
 examples/client/credentials_exchange/credentials_exchange
+examples/client/credentials_exchange/basic/basic
+examples/client/credentials_exchange/multi/multi
+examples/client/credentials_exchange/advanced/advanced
+examples/client/credentials_exchange/discovery/discovery
 examples/client/credentials_exchange/credentials_storage
+examples/client/credentials_exchange/simple_issuer_storage
+examples/client/credentials_exchange/simple_holder_storage
+examples/client/credentials_exchange/basic/basic_issuer_storage
+examples/client/credentials_exchange/basic/basic_holder_storage
+examples/client/credentials_exchange/multi/multi_issuer_storage
+examples/client/credentials_exchange/multi/multi_holder_storage
+examples/client/credentials_exchange/advanced/advanced_issuer_storage
+examples/client/credentials_exchange/advanced/advanced_holder_storage
+examples/client/credentials_exchange/discovery/discovery_issuer_storage
+examples/client/credentials_exchange/discovery/discovery_holder_storage
 examples/client/discovery_subscription/discovery_subscription
 examples/client/discovery_subscription/discovery_storage
 examples/client/group_chat/group_chat
