+++ examples/README.md
@@ -1,0 +1,71 @@
+# Self SDK Examples
+
+This directory contains comprehensive examples demonstrating the Self SDK's capabilities across different use cases and complexity levels.
+
+## 📁 Example Categories
+
+### 🎯 [Client Examples](client/) - **Recommended Starting Point**
+Modern, educational examples with progressive complexity:
+
+- **[Credential Issuance](client/credential_issuance/)** - Create and sign verifiable credentials
+- **[Credential Exchange](client/credentials_exchange/)** - Request and share credentials  
+- **[Simple Chat](client/simple_chat/)** - Basic secure messaging
+- **[Group Chat](client/group_chat/)** - Multi-party communication
+- **[Discovery Subscription](client/discovery_subscription/)** - Peer discovery
+- **[Advanced Features](client/advanced_features/)** - Production patterns
+
+### 🔧 Legacy Examples
+These examples demonstrate specific SDK features but may be less educational:
+
+- **[Chat](chat/)** - Basic messaging implementation
+- **[Agreement](agreement/)** - Agreement protocols
+- **[Discovery](discovery/)** - Peer discovery mechanisms
+- **[Credentials](credentials/)** - Credential operations
+- **[Utils](utils/)** - Utility functions and helpers
+
+## 🚀 Quick Start
+
+**New to Self SDK?** Start here:
+
+1. **[Client Examples](client/)** - Begin with the modern, educational examples
+2. Choose **[Credential Issuance](client/credential_issuance/)** or **[Credential Exchange](client/credentials_exchange/)** based on your needs
+3. Follow the progressive learning path within each example
+
+**Looking for specific features?** Browse the legacy examples for targeted implementations.
+
+## 🎓 Learning Approach
+
+### For Beginners
+- Start with **[Client Examples](client/)**
+- Follow the complexity ratings (🟢 → 🟡 → 🟠 → 🔴)
+- Complete credential workflow: Issuance → Exchange
+
+### For Experienced Developers
+- Review **[Client Examples](client/)** for modern patterns
+- Explore legacy examples for specific feature implementations
+- Use **[Advanced Features](client/advanced_features/)** for production guidance
+
+## 📚 Documentation
+
+Each example category includes:
+- **Comprehensive READMEs** with learning outcomes
+- **Progressive complexity** from beginner to expert
+- **Real-world scenarios** and practical use cases
+- **Standalone operation** - run examples independently
+
+## 🔧 Prerequisites
+
+- Go 1.19 or later
+- Self SDK dependencies (handled automatically by examples)
+
+## 🤝 Contributing
+
+When adding new examples:
+- Follow the educational patterns established in **[Client Examples](client/)**
+- Include comprehensive documentation and learning outcomes
+- Provide progressive complexity where appropriate
+- Ensure examples can run independently
+
+---
+
+**Start your Self SDK journey with [Client Examples](client/)! 🚀**
