+++ examples/client/README.md
@@ -1,0 +1,84 @@
+# Self SDK Client Examples
+
+This directory contains comprehensive examples demonstrating various Self SDK client capabilities. Each example is designed to teach specific concepts and can be used as a foundation for building your own applications.
+
+## 🎯 Quick Navigation
+
+| Example | Focus | Complexity | Best For |
+|---------|-------|------------|----------|
+| **[Credential Issuance](credential_issuance/)** | Creating & signing credentials | 🟢 Beginner-friendly | Learning credential creation |
+| **[Credential Exchange](credentials_exchange/)** | Requesting & sharing credentials | 🟢 Beginner-friendly | Learning credential workflows |
+| **[Simple Chat](simple_chat/)** | Basic messaging | 🟡 Intermediate | Understanding messaging |
+| **[Group Chat](group_chat/)** | Multi-party messaging | 🟠 Advanced | Group communication |
+| **[Discovery Subscription](discovery_subscription/)** | Peer discovery | 🟠 Advanced | Connection management |
+| **[Advanced Features](advanced_features/)** | Production patterns | 🔴 Expert | Advanced integrations |
+
+## 🚀 Getting Started
+
+### For Beginners
+Start with these examples to learn the fundamentals:
+
+1. **[Credential Issuance](credential_issuance/)** - Learn how to create and sign verifiable credentials
+2. **[Credential Exchange](credentials_exchange/)** - Learn how to request and share credentials
+
+### For Intermediate Users
+Once comfortable with basics, explore:
+
+3. **[Simple Chat](simple_chat/)** - Understand secure messaging
+4. **[Discovery Subscription](discovery_subscription/)** - Learn peer discovery
+
+### For Advanced Users
+For production-ready patterns:
+
+5. **[Group Chat](group_chat/)** - Multi-party communication
+6. **[Advanced Features](advanced_features/)** - Complex integrations
+
+## 🎓 Learning Path
+
+### Complete Credential Workflow
+For a comprehensive understanding of verifiable credentials:
+
+```
+credential_issuance/ → credentials_exchange/
+     (Create)              (Share)
+```
+
+This combination teaches the complete credential lifecycle from creation to sharing.
+
+### Communication Features
+For messaging and discovery capabilities:
+
+```
+simple_chat/ → group_chat/ → discovery_subscription/
+  (1-to-1)      (Groups)        (Discovery)
+```
+
+## 🔧 Prerequisites
+
+All examples require:
+- Go 1.19 or later
+- Self SDK dependencies (handled automatically)
+
+## 📚 Educational Features
+
+Each example directory includes:
+- **Progressive complexity** - Examples build from simple to advanced
+- **Comprehensive documentation** - Detailed explanations and learning outcomes
+- **Standalone operation** - Each example can be run independently
+- **Real-world scenarios** - Practical use cases and patterns
+
+## 🤝 Contributing
+
+Found ways to improve these examples? Contributions are welcome! Please follow the existing educational patterns and documentation standards.
+
+## 📖 Additional Resources
+
+- [Self SDK Documentation](https://docs.joinself.com)
+- [W3C Verifiable Credentials](https://w3.org/TR/vc-data-model/)
+- [Decentralized Identity Foundation](https://identity.foundation)
+
+---
+
+**Happy learning! 🎉**
+
+Choose an example that matches your current knowledge level and goals. Each example is designed to be educational and practical.
