+++ examples/client/credential_issuance/basic/go.mod
@@ -1,0 +1,9 @@
+module credential_issuance
+
+go 1.22
+
+toolchain go1.24.2
+
+replace github.com/joinself/self-go-sdk => ../../../..
+
+require github.com/joinself/self-go-sdk v0.0.0-00010101000000-000000000000
